requirements.txt
python-dotenv
pandas
tqdm
numpy
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-experimental
google-generativeai
chromadb
rank_bm25              
sentence-transformers  
agentic-doc
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
streamlit